Add tray icon and option to change startup
Added an tray icon which can be used to see if the app is running. Can close / restart control and open settings from there.

Added an option to the settings menu which can enable the app to run at system startup.
This option has been changed to false by default.